+++ package.json
@@ -18,7 +18,6 @@
     "react-dom": "^18.1.0",
     "react-hook-form": "^7.43.1",
     "react-icons": "^4.7.1",
-    "react-id-generator": "^3.0.2",
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.8.1",
     "react-scripts": "5.0.1",
