+++ src/helpers/getFriends.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-
+axios.defaults.baseURL = 'https://arcane-everglades-20653.herokuapp.com';
 export const getPartners = () => {
-  return axios.get("/partners");
+  return axios.get("/sponsors");
 };
